web/requirements.txt
redis
paho-mqtt
pandas
taospy
pyodbc
Flask